+++ src/SavingsAccount.java
@@ -1,6 +1,6 @@
 public class SavingsAccount extends AccountInfo {
 
-    public SavingsAccount(String firstName, String lastName, String emailAddress, long accountNo) {
+    public SavingsAccount(String firstName, String lastName, String emailAddress, int accountNo) {
         super(firstName, lastName, emailAddress, accountNo);
     }
 }
